# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "My API",
      "version": "1.0.0"
    },
    "paths": {
      "/users/{id}": {
        "get": {
          "summary": "get recipes associated with given user",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "integer"
              },
              "description": "The user ID"
            }
          ],
          "responses": {
            "200": {
              "description": "The list of recipes"
            },
            "400": {
              "description": "Error message"
            }
          }
        }
      },
      "/users/signup": {
        "post": {
          "summary": "signup user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "The ID of the new user"
            },
            "400": {
              "description": "Error message"
            }
          }
        }
      },
...
    }
  }